Add files via upload
This question was done for the requirement of COMP 5481 as a part of the lab assignment. This Binary tree question includes all the major concepts of Binary tree such as creating BT from String, level order traversal, searching a particular target in BT, finding distance between any 2 nodes of a BT and finding Lowest Common Ancestor between 2 nodes.