Binary tree- major concepts
This problem was done for the requirement of COMP 5481 at Concordia University as a lab assignment. This includes all the major binary tree problems condensed into 1. The concepts applied are to create the BT from the string, printing level order traversal of BT, finding distance between any 2 nodes in the tree, finding the Least common ancestor between any 2 nodes in the tree, searching for a particular target in a tree and printing elements occurring at a certain distance from the target sorted levelorder wise.