Finding smallest word combination
This problem as done as a requirement of the course COMP 5481 at Concordia University. This question seeks to find the smallest combination of consecutive letters in a particular string that contains all the letters of another string irrespective of the order. The naive approach of finding all the combinations and then printing the smallest one was ignored and a more optimized way of solving the problem was achieved.